+++ client/src/main/kotlin/io/spine/examples/chatspn/desktop/navigation/SearchField.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.examples.chatspn.desktop
+package io.spine.examples.chatspn.desktop.navigation
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -49,6 +49,7 @@ import androidx.compose.ui.input.key.Key
 import androidx.compose.ui.input.key.key
 import androidx.compose.ui.input.key.onPreviewKeyEvent
 import androidx.compose.ui.unit.dp
+import io.spine.examples.chatspn.desktop.component.IconButton
 
 /**
  * Displays the input field for searching.
@@ -138,4 +139,3 @@ private fun DecoratedTextField(
         }
     }
 }
-
